app/backend/requirements.txt
pathway
llm_app
requests
datetime
python-dotenv
kafka